+++ examples/svelte-app/src/App.svelte
@@ -34,11 +34,13 @@
         <a href="#/collapse">Collapse</a>
         <a href="#/dropdown">Dropdown</a>
         <a href="#/form">Form</a>
+        <a href="#/forward">Forward</a>
         <a href="#/icon">Icon</a>
         <a href="#/input">Input</a>
         <a href="#/modal">Modal</a>
         <a href="#/navscroll">Navscroll</a>
         <a href="#/overlay">Overlay</a>
+        <a href="#/pagination">Pagination</a>
         <a href="#/popover">Popover</a>
         <a href="#/radio">Radio</a>
         <a href="#/rating">Rating</a>