+++ examples/vanilla-app/src/main.js
@@ -39,6 +39,9 @@ document.querySelector('#app-header').innerHTML = `
         <a href="/src/affix">
           Affix
         </a>
+        <a href="/src/alert">
+          Alert
+        </a>
         <a href="/src/autocomplete">
           Autocomplete
         </a>
@@ -78,6 +81,9 @@ document.querySelector('#app-header').innerHTML = `
         <a href="/src/form">
           Form
         </a>
+        <a href="/src/forward">
+          Forward
+        </a>
         <a href="/src/icon">
           Icon
         </a>
@@ -92,6 +98,9 @@ document.querySelector('#app-header').innerHTML = `
         </a>
         <a href="/src/overlay">
           Overlay
+        </a>
+        <a href="/src/pagination.html">
+          Pagination
         </a>
         <a href="/src/popover">
           Popover